refactor: move file from componentsfolder  to modules folder
